v3.0.0: Hard fork — security hardening, modernization, namespace rename
Hard fork from AdysTech/CredentialManager. Upstream MIT attribution preserved.

Security:
- Replace BinaryFormatter with System.Text.Json (CWE-502)
- Fix persistence hardcoded to Enterprise (#69), default to LocalMachine
- JIT-safe memory zeroing via RtlZeroMemory P/Invoke
- Fix P/Invoke SetLastError, buffer size constants

Modernization:
- Rename namespace/package to shakeyourbunny.CredentialManager
- New signing key, new solution GUIDs
- Drop net45, add net8.0, keep netstandard2.0
- C# 12, nullable reference types, file-scoped namespaces
- i18n resource files (en/de/fr/es/it)
- Remove upstream AppVeyor CI config

Tests:
- Persistence parameter, JSON round-trips, null parameter handling
- Test category renamed from AppVeyor to CI
