+++ .editorconfig
@@ -1,7 +1,18 @@
-﻿[*.cs]
+root = true
 
-# IDE0008: Use explicit type
-csharp_style_var_elsewhere = false:suggestion
+[*]
+indent_style = space
+indent_size = 4
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
 
-# IDE0018: Inline variable declaration
-csharp_style_inlined_variable_declaration = false:suggestion
+[*.{xml,json,yml,yaml,resx,props,targets,csproj}]
+indent_size = 2
+
+[*.cs]
+csharp_style_namespace_declarations = file_scoped:warning
+csharp_style_var_for_built_in_types = false:suggestion
+csharp_style_var_when_type_is_apparent = true:suggestion
+csharp_prefer_braces = when_multiline:suggestion
+dotnet_sort_system_directives_first = true
