+++ .gitignore
@@ -1,192 +1,165 @@
-## Ignore Visual Studio temporary files, build results, and
-## files generated by popular Visual Studio add-ons.
-
-# User-specific files
-*.suo
-*.user
-*.sln.docstates
-
-# Build results
+## Build output
 [Dd]ebug/
 [Dd]ebugPublic/
 [Rr]elease/
 x64/
+x86/
 build/
 bld/
 [Bb]in/
 [Oo]bj/
+[Oo]ut/
+artifacts/
 
-# Roslyn cache directories
+## NuGet
+packages/
+!packages/build/
+*.nupkg
+*.snupkg
+**/[Pp]ackages/NuGet.Config
+
+## .NET / MSBuild
+*.props.user
+project.lock.json
+project.fragment.lock.json
+*.nuget.props
+*.nuget.targets
+msbuild.log
+msbuild.err
+msbuild.wrn
+
+## Roslyn
 *.ide/
+*.editorconfig.user
+.vs/
+*.GeneratedMSBuildEditorConfig.editorconfig
 
-# MSTest test Results
+## Roslyn analyzers / security audit
+*.sarif
+*.ruleset.user
+SecurityCodeScan.config.user
+
+## Test results
 [Tt]est[Rr]esult*/
 [Bb]uild[Ll]og.*
-
-#NUNIT
-*.VisualState.xml
+*.trx
+*.coverage
+*.coveragexml
 TestResult.xml
+*.VisualState.xml
 
-# Build Results of an ATL Project
-[Dd]ebugPS/
-[Rr]eleasePS/
-dlldata.c
+## Code coverage
+*.dotCover
+coverage/
+lcov.info
+*.opencover.xml
 
-*_i.c
-*_p.c
-*_i.h
-*.ilk
-*.meta
-*.obj
-*.pch
-*.pdb
-*.pgc
-*.pgd
-*.rsp
-*.sbr
-*.tlb
-*.tli
-*.tlh
-*.tmp
-*.tmp_proj
-*.log
-*.vspscc
-*.vssscc
-.builds
-*.pidb
-*.svclog
-*.scc
+## User-specific files
+*.suo
+*.user
+*.userosscache
+*.sln.docstates
 
-# Chutzpah Test files
-_Chutzpah*
-
-# Visual C++ cache files
-ipch/
+## Visual Studio
+*.rsuser
 *.aps
 *.ncb
 *.opensdf
 *.sdf
 *.cachefile
-
-# Visual Studio profiler
 *.psess
 *.vsp
 *.vspx
+*.sap
+*.[Pp]ublish.xml
+*.azurePubxml
+*.pubxml
+PublishScripts/
 
-# TFS 2012 Local Workspace
-$tf/
+## VS Code
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+*.code-workspace
 
-# Guidance Automation Toolkit
-*.gpState
+## JetBrains (Rider, PyCharm, IntelliJ)
+.idea/
+*.sln.iml
+*.iml
+*.ipr
+*.iws
+.idea_modules/
 
-# ReSharper is a .NET coding add-in
+## ReSharper
 _ReSharper*/
 *.[Rr]e[Ss]harper
 *.DotSettings.user
 
-# JustCode is a .NET coding addin-in
-.JustCode
-
-# TeamCity is a build add-in
-_TeamCity*
-
-# DotCover is a Code Coverage Tool
-*.dotCover
-
-# NCrunch
+## NCrunch
 _NCrunch_*
 .*crunch*.local.xml
 
-# MightyMoose
-*.mm.*
-AutoTest.Net/
+## TeamCity
+_TeamCity*
 
-# Web workbench (sass)
-.sass-cache/
-
-# Installshield output folder
-[Ee]xpress/
-
-# DocProject is a documentation generator add-in
-DocProject/buildhelp/
-DocProject/Help/*.HxT
-DocProject/Help/*.HxC
-DocProject/Help/*.hhc
-DocProject/Help/*.hhk
-DocProject/Help/*.hhp
-DocProject/Help/Html2
-DocProject/Help/html
-
-# Click-Once directory
-publish/
-
-# Publish Web Output
-*.[Pp]ublish.xml
-*.azurePubxml
-## TODO: Comment the next line if you want to checkin your
-## web deploy settings but do note that will include unencrypted
-## passwords
-#*.pubxml
-
-# NuGet Packages Directory
-packages/*
-## TODO: If the tool you use requires repositories.config
-## uncomment the next line
-#!packages/repositories.config
-
-# Enable "build/" folder in the NuGet Packages folder since
-# NuGet packages use it for MSBuild targets.
-# This line needs to be after the ignore of the build folder
-# (and the packages folder if the line above has been uncommented)
-!packages/build/
-
-# Windows Azure Build Output
-csx/
-*.build.csdef
-
-# Windows Store app package directory
-AppPackages/
-
-# Others
-sql/
-*.Cache
-ClientBin/
-[Ss]tyle[Cc]op.*
+## Windows
+Thumbs.db
+ehthumbs.db
+Desktop.ini
+$RECYCLE.BIN/
 ~$*
 *~
-*.dbmdl
-*.dbproj.schemaview
+
+## macOS
+.DS_Store
+.AppleDouble
+.LSOverride
+._*
+
+## Common temporary and backup files
+*.bak
+*.bak2
+*.orig
+*.tmp
+*.tmp_proj
+*.swp
+*.swo
+*~
+*.log
+*.old
+
+## Native build artifacts
+*.obj
+*.pch
+*.pdb
+*.pgc
+*.pgd
+*.ilk
+*.meta
+*.tlb
+*.tli
+*.tlh
+*.sbr
+*.rsp
+dlldata.c
+*_i.c
+*_p.c
+*_i.h
+
+## Secrets and credentials (never commit these)
 *.pfx
 *.publishsettings
-node_modules/
+*.key
+!key.snk
 
-# RIA/Silverlight projects
-Generated_Code/
-
-# Backup & report files from converting an old project file
-# to a newer Visual Studio version. Backup files are not needed,
-# because we have git ;-)
-_UpgradeReport_Files/
-Backup*/
-UpgradeLog*.XML
-UpgradeLog*.htm
-
-# SQL Server files
+## SQL Server
 *.mdf
 *.ldf
 
-# Business Intelligence projects
-*.rdl.data
-*.bim.layout
-*.bim_*.settings
-
-# Microsoft Fakes
+## Microsoft Fakes
 FakesAssemblies/
 
-# LightSwitch generated files
-GeneratedArtifacts/
-_Pvt_Extensions/
-ModelManifest.xml
-/.vs
-/.vs
-/.vscode/tasks.json
+## NUnit
+nunit-*.xml
