+++ CredentialManager.sln
@@ -3,22 +3,21 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29503.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdysTech.CredentialManager", "src\AdysTech.CredentialManager\AdysTech.CredentialManager.csproj", "{ABA77D40-8D43-42EE-BE03-44E4AD93672A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "shakeyourbunny.CredentialManager", "src\shakeyourbunny.CredentialManager\shakeyourbunny.CredentialManager.csproj", "{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CredentialManager.Test", "tests\CredentialManager.Test\CredentialManager.Test.csproj", "{5DCA2C07-C1F3-424F-B934-F952595586CE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CredentialManager.Test", "tests\CredentialManager.Test\CredentialManager.Test.csproj", "{18A3699C-31F6-400F-8935-86E65E5B36E5}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{EBCE1FF4-81EB-4EC9-A085-2CA0F81100E5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{835DA67E-EA6C-4E05-8C2E-4601CDE65EA1}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
-		appveyor.yml = appveyor.yml
-		changelog.md = changelog.md
+		CHANGELOG.md = CHANGELOG.md
 		Directory.Build.props = Directory.Build.props
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0609161C-CF52-44B3-A09E-6CCD4BCCB30F}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{63C41070-5FDD-4CB0-8046-AF6617DFF800}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{7E19D2AF-C0F0-4229-8E95-5B8ADC0BE732}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{3D02AC64-9CEB-4161-83C3-2AA58A0821CC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,31 +27,31 @@ Global
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Debug|x64.Build.0 = Debug|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Release|x64.ActiveCfg = Release|Any CPU
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A}.Release|x64.Build.0 = Release|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Debug|x64.Build.0 = Debug|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.ActiveCfg = Release|Any CPU
-		{5DCA2C07-C1F3-424F-B934-F952595586CE}.Release|x64.Build.0 = Release|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Debug|x64.Build.0 = Debug|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Release|x64.ActiveCfg = Release|Any CPU
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D}.Release|x64.Build.0 = Release|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Debug|x64.Build.0 = Debug|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Release|x64.ActiveCfg = Release|Any CPU
+		{18A3699C-31F6-400F-8935-86E65E5B36E5}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{ABA77D40-8D43-42EE-BE03-44E4AD93672A} = {0609161C-CF52-44B3-A09E-6CCD4BCCB30F}
-		{5DCA2C07-C1F3-424F-B934-F952595586CE} = {7E19D2AF-C0F0-4229-8E95-5B8ADC0BE732}
+		{4DD8348E-AAD9-47DA-AF28-5BDF5FDF431D} = {63C41070-5FDD-4CB0-8046-AF6617DFF800}
+		{18A3699C-31F6-400F-8935-86E65E5B36E5} = {3D02AC64-9CEB-4161-83C3-2AA58A0821CC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {36A3FDA1-ACC5-4934-8E8A-DBE3ACADFE39}
+		SolutionGuid = {D81D02DD-15CE-4066-B904-E3F5F469427F}
 	EndGlobalSection
 EndGlobal
