+++ Directory.Build.props
@@ -1,16 +1,16 @@
 <Project>
- <PropertyGroup>
-   <Deterministic>true</Deterministic>
-   <TargetFrameworks>net45;netstandard2.0</TargetFrameworks>
-   <LangVersion>8</LangVersion>
-   <PackageProjectUrl>https://github.com/AdysTech/CredentialManager</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/AdysTech/CredentialManager</RepositoryUrl>
+  <PropertyGroup>
+    <Deterministic>true</Deterministic>
+    <TargetFrameworks>netstandard2.0;net8.0</TargetFrameworks>
+    <LangVersion>12.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <PackageProjectUrl>https://github.com/shakeyourbunny/CredentialManager</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/shakeyourbunny/CredentialManager</RepositoryUrl>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Copyright>© AdysTech 2016-2022</Copyright>
-    <Authors>AdysTech</Authors>
+    <Copyright>© Adys Tech 2016-2022, © shakeyourbunny 2026</Copyright>
+    <Authors>AdysTech;shakeyourbunny</Authors>
     <SignAssembly>true</SignAssembly>
-    <DelaySign>false</DelaySign> 
-    <SignAssembly>true</SignAssembly>
+    <DelaySign>false</DelaySign>
     <AssemblyOriginatorKeyFile>../../key.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
 </Project>
