+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
 Copyright (c) 2018 Adys Tech
+Copyright (c) 2026 shakeyourbunny <shakeyourbunny@gmx.net>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
