--- appveyor.yml
@@ -1,31 +1,0 @@
-version: 0.{build}
-image: Visual Studio 2019
-pull_requests:
-  do_not_increment_build_number: true
-branches:
-  only:
-  - master
-before_build:
-- nuget restore
-- dotnet restore
-configuration:
-  - Release
-build:
-  verbosity: minimal
-  # publish_nuget: true
-  publish_nuget_symbols: true
-  use_snupkg_format: true
-test:
-  categories:
-    only:
-      - AppVeyor
-deploy:
-- provider: NuGet
-  api_key:
-    secure: z+VCuAl52RiKf1Ie2p6ivQloYTH12q7H0G6fEl/c4FepMD75/vI7oRnbMISdgWEm
-  on:
-    branch: master
-artifacts:
-- path: src\AdysTech.CredentialManager\bin\Debug
-  name: CredentialManager
-- path: src\AdysTech.CredentialManager\bin\**\*nupkg
