--- src/AdysTech.CredentialManager/AdysTech.CredentialManager.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-<!-- Rest of the properies are taken from build.props, these are mostly for the nuget package -->
-  <PropertyGroup>
-    <Description>
-      C# wrapper around CredWrite / CredRead functions to store and retrieve from Windows Credential Store.
-      The library .NET Standard 1.2+ and .NET Framework 4.5+ and provides
-      1. Prompt user for Windows Generic Credentials
-      2. Save Windows Generic Credentials
-      3. Retrieve saved Windows Generic Credentials
-      4. Add or edit ccomments, store additional attribute objects associated with Credentials</Description> 
-    <Version>2.6.0.0</Version>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-    <Copyright>© AdysTech 2016-2022</Copyright>
-    <AssemblyVersion>2.6.0.0</AssemblyVersion>
-    <FileVersion>2.6.0.0</FileVersion>
-  </PropertyGroup>
-</Project>
