+++ src/shakeyourbunny.CredentialManager/Properties/AssemblyInfo.cs
