+++ src/shakeyourbunny.CredentialManager/Resources/Strings.de.resx
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="utf-8"?>
+<root>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
+    <xsd:element name="root" msdata:IsDataSet="true">
+      <xsd:complexType>
+        <xsd:choice maxOccurs="unbounded">
+          <xsd:element name="metadata">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
+              </xsd:sequence>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="assembly">
+            <xsd:complexType>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="data">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="resheader">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
+            </xsd:complexType>
+          </xsd:element>
+        </xsd:choice>
+      </xsd:complexType>
+    </xsd:element>
+  </xsd:schema>
+  <resheader name="resmimetype">
+    <value>text/microsoft-resx</value>
+  </resheader>
+  <resheader name="version">
+    <value>2.0</value>
+  </resheader>
+  <resheader name="reader">
+    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <resheader name="writer">
+    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <data name="CommentTooLong" xml:space="preserve">
+    <value>Kommentar darf nicht mehr als 256 Bytes lang sein</value>
+  </data>
+  <data name="TargetNameNullOrEmpty" xml:space="preserve">
+    <value>TargetName darf nicht null oder leer sein</value>
+  </data>
+  <data name="TargetNameTooLong" xml:space="preserve">
+    <value>TargetName darf nicht mehr als 32 kB lang sein</value>
+  </data>
+  <data name="CredentialBlobNullOrEmpty" xml:space="preserve">
+    <value>CredentialBlob darf nicht null oder leer sein</value>
+  </data>
+  <data name="CredentialBlobTooLong" xml:space="preserve">
+    <value>Anmeldedaten dürfen nicht mehr als {0} Bytes lang sein</value>
+  </data>
+  <data name="TooManyAttributes" xml:space="preserve">
+    <value>Anmeldedaten dürfen nicht mehr als 64 Attribute haben</value>
+  </data>
+  <data name="AttributeNameTooLong" xml:space="preserve">
+    <value>Attributnamen dürfen nicht mehr als 256 Bytes lang sein. Fehler bei Schlüssel: {0}</value>
+  </data>
+  <data name="AttributeValueNull" xml:space="preserve">
+    <value>Attributwert darf nicht null sein. Fehler bei Schlüssel: {0}</value>
+  </data>
+  <data name="AttributeValueTooLong" xml:space="preserve">
+    <value>Attributwerte dürfen nach der Serialisierung nicht mehr als 256 Bytes lang sein. Fehler bei Wert für Schlüssel: {0}</value>
+  </data>
+  <data name="UnableToSaveCredential" xml:space="preserve">
+    <value>Anmeldedaten konnten nicht gespeichert werden</value>
+  </data>
+  <data name="UnableToParseUserName" xml:space="preserve">
+    <value>Benutzername konnte nicht analysiert werden</value>
+  </data>
+  <data name="UnableToDeleteCredential" xml:space="preserve">
+    <value>Anmeldedaten konnten nicht gelöscht werden</value>
+  </data>
+  <data name="UnableToEnumerateCredentials" xml:space="preserve">
+    <value>Anmeldedatenspeicher konnte nicht aufgelistet werden</value>
+  </data>
+  <data name="UnableToReadCredential" xml:space="preserve">
+    <value>Anmeldedaten konnten nicht gelesen werden</value>
+  </data>
+  <data name="InvalidCriticalHandle" xml:space="preserve">
+    <value>Ungültiges CriticalHandle!</value>
+  </data>
+</root>
