+++ src/shakeyourbunny.CredentialManager/Resources/Strings.fr.resx
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="utf-8"?>
+<root>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
+    <xsd:element name="root" msdata:IsDataSet="true">
+      <xsd:complexType>
+        <xsd:choice maxOccurs="unbounded">
+          <xsd:element name="metadata">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
+              </xsd:sequence>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="assembly">
+            <xsd:complexType>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="data">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="resheader">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
+            </xsd:complexType>
+          </xsd:element>
+        </xsd:choice>
+      </xsd:complexType>
+    </xsd:element>
+  </xsd:schema>
+  <resheader name="resmimetype">
+    <value>text/microsoft-resx</value>
+  </resheader>
+  <resheader name="version">
+    <value>2.0</value>
+  </resheader>
+  <resheader name="reader">
+    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <resheader name="writer">
+    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <data name="CommentTooLong" xml:space="preserve">
+    <value>Le commentaire ne peut pas dépasser 256 octets</value>
+  </data>
+  <data name="TargetNameNullOrEmpty" xml:space="preserve">
+    <value>TargetName ne peut pas être nul ou vide</value>
+  </data>
+  <data name="TargetNameTooLong" xml:space="preserve">
+    <value>TargetName ne peut pas dépasser 32 ko</value>
+  </data>
+  <data name="CredentialBlobNullOrEmpty" xml:space="preserve">
+    <value>CredentialBlob ne peut pas être nul ou vide</value>
+  </data>
+  <data name="CredentialBlobTooLong" xml:space="preserve">
+    <value>Les informations d'identification ne peuvent pas dépasser {0} octets</value>
+  </data>
+  <data name="TooManyAttributes" xml:space="preserve">
+    <value>Les informations d'identification ne peuvent pas avoir plus de 64 attributs</value>
+  </data>
+  <data name="AttributeNameTooLong" xml:space="preserve">
+    <value>Les noms d'attribut ne peuvent pas dépasser 256 octets. Erreur avec la clé : {0}</value>
+  </data>
+  <data name="AttributeValueNull" xml:space="preserve">
+    <value>La valeur de l'attribut ne peut pas être nulle. Erreur avec la clé : {0}</value>
+  </data>
+  <data name="AttributeValueTooLong" xml:space="preserve">
+    <value>Les valeurs d'attribut ne peuvent pas dépasser 256 octets après sérialisation. Erreur avec la valeur pour la clé : {0}</value>
+  </data>
+  <data name="UnableToSaveCredential" xml:space="preserve">
+    <value>Impossible d'enregistrer les informations d'identification</value>
+  </data>
+  <data name="UnableToParseUserName" xml:space="preserve">
+    <value>Impossible d'analyser le nom d'utilisateur</value>
+  </data>
+  <data name="UnableToDeleteCredential" xml:space="preserve">
+    <value>Impossible de supprimer les informations d'identification</value>
+  </data>
+  <data name="UnableToEnumerateCredentials" xml:space="preserve">
+    <value>Impossible d'énumérer le magasin d'informations d'identification</value>
+  </data>
+  <data name="UnableToReadCredential" xml:space="preserve">
+    <value>Impossible de lire les informations d'identification</value>
+  </data>
+  <data name="InvalidCriticalHandle" xml:space="preserve">
+    <value>CriticalHandle non valide !</value>
+  </data>
+</root>
