+++ src/shakeyourbunny.CredentialManager/Resources/Strings.it.resx
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="utf-8"?>
+<root>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
+    <xsd:element name="root" msdata:IsDataSet="true">
+      <xsd:complexType>
+        <xsd:choice maxOccurs="unbounded">
+          <xsd:element name="metadata">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
+              </xsd:sequence>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="assembly">
+            <xsd:complexType>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="data">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
+            </xsd:complexType>
+          </xsd:element>
+          <xsd:element name="resheader">
+            <xsd:complexType>
+              <xsd:sequence>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+              </xsd:sequence>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
+            </xsd:complexType>
+          </xsd:element>
+        </xsd:choice>
+      </xsd:complexType>
+    </xsd:element>
+  </xsd:schema>
+  <resheader name="resmimetype">
+    <value>text/microsoft-resx</value>
+  </resheader>
+  <resheader name="version">
+    <value>2.0</value>
+  </resheader>
+  <resheader name="reader">
+    <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <resheader name="writer">
+    <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </resheader>
+  <data name="CommentTooLong" xml:space="preserve">
+    <value>Il commento non può superare i 256 byte</value>
+  </data>
+  <data name="TargetNameNullOrEmpty" xml:space="preserve">
+    <value>TargetName non può essere nullo o vuoto</value>
+  </data>
+  <data name="TargetNameTooLong" xml:space="preserve">
+    <value>TargetName non può superare i 32 kB</value>
+  </data>
+  <data name="CredentialBlobNullOrEmpty" xml:space="preserve">
+    <value>CredentialBlob non può essere nullo o vuoto</value>
+  </data>
+  <data name="CredentialBlobTooLong" xml:space="preserve">
+    <value>Le credenziali non possono superare i {0} byte</value>
+  </data>
+  <data name="TooManyAttributes" xml:space="preserve">
+    <value>Le credenziali non possono avere più di 64 attributi</value>
+  </data>
+  <data name="AttributeNameTooLong" xml:space="preserve">
+    <value>I nomi degli attributi non possono superare i 256 byte. Errore con la chiave: {0}</value>
+  </data>
+  <data name="AttributeValueNull" xml:space="preserve">
+    <value>Il valore dell'attributo non può essere nullo. Errore con la chiave: {0}</value>
+  </data>
+  <data name="AttributeValueTooLong" xml:space="preserve">
+    <value>I valori degli attributi non possono superare i 256 byte dopo la serializzazione. Errore con il valore per la chiave: {0}</value>
+  </data>
+  <data name="UnableToSaveCredential" xml:space="preserve">
+    <value>Impossibile salvare le credenziali</value>
+  </data>
+  <data name="UnableToParseUserName" xml:space="preserve">
+    <value>Impossibile analizzare il nome utente</value>
+  </data>
+  <data name="UnableToDeleteCredential" xml:space="preserve">
+    <value>Impossibile eliminare le credenziali</value>
+  </data>
+  <data name="UnableToEnumerateCredentials" xml:space="preserve">
+    <value>Impossibile enumerare l'archivio delle credenziali</value>
+  </data>
+  <data name="UnableToReadCredential" xml:space="preserve">
+    <value>Impossibile leggere le credenziali</value>
+  </data>
+  <data name="InvalidCriticalHandle" xml:space="preserve">
+    <value>CriticalHandle non valido!</value>
+  </data>
+</root>
