+++ src/shakeyourbunny.CredentialManager/shakeyourbunny.CredentialManager.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <Description>
+      C# wrapper around CredWrite / CredRead functions to store and retrieve from Windows Credential Store.
+      Supports .NET 8.0+ and .NET Standard 2.0+. Provides credential prompting, save/retrieve,
+      enumeration, and secure attribute storage using JSON serialization.</Description>
+    <Version>3.0.0</Version>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <IncludeSymbols>true</IncludeSymbols>
+    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
+    <AssemblyVersion>3.0.0.0</AssemblyVersion>
+    <FileVersion>3.0.0.0</FileVersion>
+  </PropertyGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="System.Text.Json" Version="8.0.5" />
+  </ItemGroup>
+</Project>
